Restructure files
